refactor: change in tree structure
