+++ shop-info-service/src/functions/schema.ts
