--- shop-info-service/src/resources/test-id.json
@@ -1,5 +1,0 @@
-{
-  "pathParameters": {
-    "id": "c80c86cf-f3c6-45e6-87c5-6f94cbe7f9f5"
-  }
-}
