+++ shop-info-service/src/test/resources/test-id.json
@@ -1,0 +1,5 @@
+{
+  "pathParameters": {
+    "id": "2f2fc36b-c0ad-476b-b1d4-7051672dcc28"
+  }
+}
