+++ shop-info-service/src/test/resources/test-post.json
