+++ shop-info-service/src/test/resources/test.json
